# api/swagger/swagger.yaml
swagger: "2.0"
info:
  title: "Event server"
  version: 1.0.0
  description: ""
basePath: /
paths:
...
  /event/{id}:
    get:
      tags:
        - Event
      summary: Find event by ID 
      operationId: findById
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of event to return
          required: true
          type: string
      responses:
        200:
          description: OK
        default:
          description: Unexpected error
    put:
      tags:
        - Event
      summary: Updated event data
      operationId: updateById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of event to return
          required: true
          type: string
      responses:
        200:
          description: Successfully updated
        404:
          description: Event has not been found
    delete:
      tags:
        - Event
      summary: Delete event from list
      operationId: deleteById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Id of event that needs to be deleted
          required: true
          type: string
      responses:
        200:
          description: Event has been successfully removed
        404:
          description: Event has not been found
...